Adicionando códigos de eventos corrigidos e modularizados, códigos das telas modularizados, correções de código...
